+++ backend/analysis/cron/parse_job.py
@@ -1,11 +1,9 @@
+from ..parse.parse import create_utterance_objects, parse_transcript
+from ..models import Transcript
+from django_cron import CronJobBase, Schedule
+from django.db.models import Q
 import logging
 import os
-
-from django.db.models import Q
-from django_cron import CronJobBase, Schedule
-
-from ..models import Transcript
-from ..parse.parse import create_utterance_objects, parse_transcript
 
 logger = logging.getLogger('sasta')
 
@@ -17,7 +15,8 @@ class ParseJob(CronJobBase):
     code = 'sasta.parse_job'  # a unique code
 
     def do(self):
-        for transcript in Transcript.objects.filter(Q(status='converted') | Q(status='parsing-failed')):  # noqa: E501
+        for transcript in Transcript.objects.filter(
+                Q(status='converted') | Q(status='parsing-failed')):
             try:
                 output_path = transcript.content.path.replace(
                     '/transcripts', '/parsed')