+++ backend/convert/chat_reader_test.py
@@ -1,7 +1,29 @@
+import os.path as op
+
+import pytest
+from analysis.conftest import stap_category, tarsp_category
+
 from .chat_reader import ChatDocument
+from .conftest import TEST_DIR
 
 
-def test_chat_reader(chafiles):
+@pytest.mark.django_db
+@pytest.mark.usefixtures("tarsp_category")
+def test_chat_reader(chafiles, tarsp_category):
     for (input, _) in chafiles:
-        doc = ChatDocument.from_chatfile(input)
+        doc = ChatDocument.from_chatfile(input, tarsp_category)
         assert len(doc.lines)
+
+
+@pytest.mark.django_db
+@pytest.mark.usefixtures("tarsp_category", "stap_category")
+def test_marking_postcodes(chafiles, tarsp_category, stap_category):
+    inf = op.join(TEST_DIR, 'TD16.cha')
+
+    tarsp_doc = ChatDocument.from_chatfile(inf, tarsp_category)
+    tarsp_num_xsid = len([x for x in tarsp_doc.lines if x.tiers.get('xsid')])
+    assert tarsp_num_xsid == 42
+
+    stap_doc = ChatDocument.from_chatfile(inf, stap_category)
+    stap_num_xsid = len([x for x in stap_doc.lines if x.tiers.get('xsid')])
+    assert stap_num_xsid == 42 + 16