+++ backend/convert/chat_writer_test.py
@@ -1,12 +1,17 @@
+import pytest
+from analysis.conftest import tarsp_category
+
 from .chat_reader import ChatDocument
 from .chat_writer import ChatWriter
 
 
-def test_chat_writer(chafiles):
+@pytest.mark.django_db
+@pytest.mark.usefixtures("tarsp_category")
+def test_chat_writer(chafiles, tarsp_category):
     for inpath, outpath in chafiles:
-        doc = ChatDocument.from_chatfile(inpath)
+        doc = ChatDocument.from_chatfile(inpath, tarsp_category)
         with open(outpath, 'w+', encoding='utf-8') as target:
             writer = ChatWriter(doc, target=target)
             writer.write()
-        out_doc = ChatDocument.from_chatfile(outpath)
+        out_doc = ChatDocument.from_chatfile(outpath, tarsp_category)
         assert out_doc == doc