+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes (beginning at version 0.2.0) to this project will be documen
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.6] - 2021-12-17
+### Changed
+- Differentiate CHAT postcode markers by method. For STAP `[+ VU]` is marked, for all methods `[+ G]`
+
 ## [0.2.5] - 2021-11-12
 ### Added
 - Process CHAT postcodes `[+ VU]` and `[+ G]` as utterance analysis markers
@@ -21,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.0] - 2021-09-10
 ### Added
 - CHAT format input.
-- Asynchronous parsing; the parse process continues in the background. 
+- Asynchronous parsing; the parse process continues in the background.
 - Updates to query definitions and related SASTADEV functions.
 - Automatic corrections on input files to improve parses of irregular language (implemented in SASTADEV).
 - Logos for participating organisations.